requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# LangChain ecosystem
langchain==0.3.1
langchain-community==0.3.1
langchain-huggingface==0.1.0
langchain-groq==0.2.0

# PDF parsing
pypdf==4.3.1

# Embeddings (local, free)
sentence-transformers==3.1.1

# Vector store
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.1
